Added checks for bad requests

+++ BudgetTrackerAPI/Services/TransactionService/TransactionService.cs
@@ -1,7 +1,4 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-
-namespace BudgetTrackerAPI.Services.TransactionService
+﻿namespace BudgetTrackerAPI.Services.TransactionService
 {
     public class TransactionService : ITransactionService
     {
@@ -19,17 +16,21 @@ namespace BudgetTrackerAPI.Services.TransactionService
             return await _context.Transactions.ToListAsync();
         }
 
-        public async Task<List<Transaction>> DeleteTransaction(int id)
+        public async Task<List<Transaction>?> DeleteTransaction(int id)
         {
             var transactionToDelete = await _context.Transactions.FindAsync(id);
+            if (transactionToDelete == null) return null;
+
             _context.Transactions.Remove(transactionToDelete);
+
             await _context.SaveChangesAsync();
             return await _context.Transactions.ToListAsync();
         }
 
-        public async Task<List<Transaction>> EditTransaction(Transaction newTransaction)
+        public async Task<List<Transaction>?> EditTransaction(Transaction newTransaction)
         {
             var transactionToEdit = await _context.Transactions.FindAsync(newTransaction.Id);
+            if (transactionToEdit == null) return null;
 
             transactionToEdit.Id = newTransaction.Id;
             transactionToEdit.Value = newTransaction.Value;
@@ -60,15 +61,9 @@ namespace BudgetTrackerAPI.Services.TransactionService
         {
             if (startDate == "" && endDate == "") return null;
 
-            if (startDate == "")
-            {
-                return await _context.Transactions.Where(t => t.Date <= DateTime.Parse(endDate)).ToListAsync();
-            }
+            if (startDate == "") return await _context.Transactions.Where(t => t.Date <= DateTime.Parse(endDate)).ToListAsync();
 
-            if (endDate == "")
-            {
-                return await _context.Transactions.Where(t => t.Date >= DateTime.Parse(startDate)).ToListAsync();
-            }
+            if (endDate == "") return await _context.Transactions.Where(t => t.Date >= DateTime.Parse(startDate)).ToListAsync();
 
             return await _context.Transactions.Where(t => t.Date >= DateTime.Parse(startDate) && t.Date <= DateTime.Parse(endDate)).ToListAsync();
         }
@@ -83,7 +78,7 @@ namespace BudgetTrackerAPI.Services.TransactionService
             return await _context.Transactions.Where(t => t.TransactionType == TransactionType.OUTGOING).ToListAsync();
         }
 
-        public async Task<Transaction> GetSingleTransactionById(int id)
+        public async Task<Transaction?> GetSingleTransactionById(int id)
         {
             var transaction = await _context.Transactions.FindAsync(id);
             return transaction;
